+++ src/WireGuard/ServerConfig.php
@@ -52,18 +52,9 @@ class ServerConfig
 
         $this->registerPublicKey($nodeNumber, $publicKey);
 
-        if (null !== $useMtu = $this->wgConfig->useMtu()) {
-            return <<< EOF
-                [Interface]
-                MTU = {$useMtu}
-                Address = {$ipList}
-                ListenPort = {$this->wgConfig->listenPort()}
-                PrivateKey = {{PRIVATE_KEY}}
-                EOF;
-        }
-
         return <<< EOF
             [Interface]
+            MTU = {$this->wgConfig->setMtu()}
             Address = {$ipList}
             ListenPort = {$this->wgConfig->listenPort()}
             PrivateKey = {{PRIVATE_KEY}}